+++ streams-tests/jvm/src/test/scala/zio/stream/StreamLazinessSpec.scala
@@ -1,4 +1,4 @@
-package zio.stream.experimental
+package zio.stream
 
 import zio._
 import zio.test._
