+++ streams-tests/shared/src/test/scala/zio/stream/SinkUtils.scala
@@ -1,4 +1,4 @@
-package zio.stream.experimental
+package zio.stream
 
 import zio.test.Assertion.equalTo
 import zio.test.{Assertion, TestResult, assert}
