+++ streams-tests/shared/src/test/scala/zio/stream/ZStreamGen.scala
@@ -1,4 +1,4 @@
-package zio.stream.experimental
+package zio.stream
 
 import zio._
 import zio.test.{Gen, GenZIO, Sized}
