+++ streams/shared/src/main/scala-2.13+/zio/stream/ZPipelineVersionSpecific.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package zio.stream.experimental
+package zio.stream
 
 import zio.stacktracer.TracingImplicits.disableAutoTrace
