+++ streams/shared/src/main/scala/zio/stream/SubscriptionRef.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package zio.stream.experimental
+package zio.stream
 
 import zio._
 import zio.stacktracer.TracingImplicits.disableAutoTrace
