+++ streams/shared/src/main/scala/zio/stream/ZChannel.scala
@@ -1,4 +1,4 @@
-package zio.stream.experimental
+package zio.stream
 
 import zio.ZManaged.ReleaseMap
 import zio._
