+++ streams/shared/src/main/scala/zio/stream/ZSink.scala
@@ -1,4 +1,4 @@
-package zio.stream.experimental
+package zio.stream
 
 import zio._
 import zio.stream.internal.CharacterSet._
