+++ streams/shared/src/main/scala/zio/stream/ZStream.scala
@@ -1,4 +1,4 @@
-package zio.stream.experimental
+package zio.stream
 
 import zio._
 import zio.internal.{SingleThreadedRingBuffer, UniqueKey}
