+++ streams/shared/src/main/scala/zio/stream/internal/Utils.scala
@@ -1,4 +1,4 @@
-package zio.stream.experimental.internal
+package zio.stream.internal
 
 import zio.Chunk
 import zio.stacktracer.TracingImplicits.disableAutoTrace
